#54 refactored athlete widget structure
