+++ frontend/lib/ui/view/tenant/athletes/athlete_details/athlete_detail_controller.dart
@@ -1,0 +1,8 @@
+
+
+import '../../../../../packages/gettools/stateful_controller.dart';
+
+class AthleteController extends StatefulGetxController {
+
+
+}
