+++ frontend/lib/ui/view/tenant/athletes/athlete_details/athlete_details_screen.dart
@@ -1,0 +1,12 @@
+
+
+import 'package:flutter/cupertino.dart';
+
+class AthleteDetailScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    // TODO: implement build
+    throw UnimplementedError();
+  }
+
+}
